fix reserve
